adding references and a little clean-up

+++ kernel/gdt.c
@@ -66,7 +66,7 @@ void gdt_install()
 {
     /* Set up GDT pointer and limit */
     gp.limit = (sizeof(struct gdt_entry) * 3) - 1;
-    gp.base = &gdt;
+    gp.base = (uint32_t)&gdt;
 
     /* NULL descriptor */
     gdt_set_gate(0, 0, 0, 0, 0);
